Factorize nifs for OTP crypto module
Move implementation from ESP32's platform_nifs.c to otp_crypto.c and make it
available to all platforms using mbedtls.

Alter Pico's mbedtls configuration to handle the ciphers we support.
