+++ src/platforms/esp32/components/avm_builtins/Kconfig
@@ -66,6 +66,10 @@ config AVM_ENABLE_UART_PORT_DRIVER
     bool "Enable UART port driver"
     default y
 
+config AVM_ENABLE_OTP_CRYPTO_NIFS
+    bool "Enable OTP Crypto NIFs"
+    default y
+
 config AVM_ENABLE_OTP_SOCKET_NIFS
     bool "Enable OTP Socket NIFs"
     default y
