+++ src/platforms/esp32/components/avm_sys/CMakeLists.txt
@@ -26,6 +26,7 @@ set(AVM_SYS_COMPONENT_SRCS
     "platform_nifs.c"
     "platform_defaultatoms.c"
     "../../../../libAtomVM/inet.c"
+    "../../../../libAtomVM/otp_crypto.c"
     "../../../../libAtomVM/otp_net.c"
     "../../../../libAtomVM/otp_socket.c"
     "../../../../libAtomVM/otp_ssl.c"
