+++ src/platforms/generic_unix/lib/CMakeLists.txt
@@ -67,21 +67,23 @@ if (MbedTLS_FOUND)
     target_compile_definitions(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC ATOMVM_HAS_MBEDTLS)
     target_sources(libAtomVM${PLATFORM_LIB_SUFFIX}
         PRIVATE
+        ../../../libAtomVM/otp_crypto.c
+        ../../../libAtomVM/otp_crypto.h
         ../../../libAtomVM/otp_ssl.c
         ../../../libAtomVM/otp_ssl.h
     )
 else()
-    message("WARNING:  Could NOT find MbedTLS, SSL will not be supported. Install MbedTLS 3.x or try to set MBEDTLS_ROOT_DIR to installation prefix of MbedTLS 2.x")
+    message("WARNING:  Could NOT find MbedTLS, ssl and crypto modules will not be supported. Install MbedTLS 3.x or try to set MBEDTLS_ROOT_DIR to installation prefix of MbedTLS 2.x")
 endif()
 
-# For now we still use OpenSSL for random and crypto
+# For now we still use OpenSSL for random
 find_package(OpenSSL)
 if (${OPENSSL_FOUND} STREQUAL TRUE)
     target_include_directories(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC ${OPENSSL_INCLUDE_DIR})
     target_compile_definitions(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC ATOMVM_HAS_OPENSSL)
     target_link_libraries(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC ${OPENSSL_CRYPTO_LIBRARY})
 else()
-    message("WARNING:  Some crypto operations will not be supported.")
+    message("WARNING:  atomvm:random/0 and atomvm:rand_bytes/1 will not be supported.")
 endif()
 
 # enable by default dynamic loading on unix
