+++ src/platforms/generic_unix/lib/platform_nifs.c
@@ -25,6 +25,7 @@
 #include "interop.h"
 #include "memory.h"
 #include "nifs.h"
+#include "otp_crypto.h"
 #include "otp_net.h"
 #include "otp_socket.h"
 #include "otp_ssl.h"
@@ -33,7 +34,6 @@
 #include <stdlib.h>
 
 #if defined ATOMVM_HAS_OPENSSL
-#include <openssl/evp.h>
 #include <openssl/rand.h>
 #endif
 
@@ -53,120 +53,6 @@
     return term_invalid_term();
 
 #if defined ATOMVM_HAS_OPENSSL
-
-enum crypto_algorithm
-{
-    CryptoInvalidAlgorithm = 0,
-    CryptoMd5,
-    CryptoSha1,
-    CryptoSha224,
-    CryptoSha256,
-    CryptoSha384,
-    CryptoSha512
-};
-
-static const AtomStringIntPair crypto_algorithm_table[] = {
-    { ATOM_STR("\x3", "md5"), CryptoMd5 },
-    { ATOM_STR("\x3", "sha"), CryptoSha1 },
-    { ATOM_STR("\x6", "sha224"), CryptoSha224 },
-    { ATOM_STR("\x6", "sha256"), CryptoSha256 },
-    { ATOM_STR("\x6", "sha384"), CryptoSha384 },
-    { ATOM_STR("\x6", "sha512"), CryptoSha512 },
-    SELECT_INT_DEFAULT(CryptoInvalidAlgorithm)
-};
-
-const char *get_crypto_algorithm(GlobalContext *global, term type)
-{
-    enum crypto_algorithm algo = interop_atom_term_select_int(crypto_algorithm_table, type, global);
-    switch (algo) {
-        case CryptoMd5:
-            return "MD5";
-        case CryptoSha1:
-            return "SHA1";
-        case CryptoSha224:
-            return "SHA224";
-        case CryptoSha256:
-            return "SHA256";
-        case CryptoSha384:
-            return "SHA384";
-        case CryptoSha512:
-            return "SHA512";
-        default:
-            return NULL;
-    }
-}
-
-static InteropFunctionResult hash_fold_fun(term t, void *accum)
-{
-    EVP_MD_CTX *md_ctx = (EVP_MD_CTX *) accum;
-    if (term_is_any_integer(t)) {
-        avm_int64_t tmp = term_maybe_unbox_int64(t);
-        if (tmp < 0 || tmp > 255) {
-            return InteropBadArg;
-        }
-        uint8_t val = (avm_int64_t) tmp;
-        if (!EVP_DigestUpdate(md_ctx, &val, 1)) {
-            return InteropBadArg;
-        }
-    } else if (term_is_binary(t)) {
-        if (!EVP_DigestUpdate(md_ctx, (uint8_t *) term_binary_data(t), term_binary_size(t))) {
-            return InteropBadArg;
-        }
-    }
-    return InteropOk;
-}
-
-static term nif_crypto_hash(Context *ctx, int argc, term argv[])
-{
-    UNUSED(argc);
-    term type = argv[0];
-    term data = argv[1];
-    VALIDATE_VALUE(type, term_is_atom);
-
-    const char *algorithm = get_crypto_algorithm(ctx->global, type);
-    if (IS_NULL_PTR(algorithm)) {
-        RAISE_ERROR(BADARG_ATOM);
-    }
-
-    const EVP_MD *md = EVP_get_digestbyname(algorithm);
-    if (IS_NULL_PTR(md)) {
-        fprintf(stderr, "Could not find digest by name: %s\n", algorithm);
-        RAISE_ERROR(BADARG_ATOM);
-    }
-    EVP_MD_CTX *md_ctx = EVP_MD_CTX_new();
-    if (IS_NULL_PTR(md_ctx)) {
-        fprintf(stderr, "Could not allocate MD context\n");
-        RAISE_ERROR(OUT_OF_MEMORY_ATOM);
-    }
-
-    if (UNLIKELY(!EVP_DigestInit_ex(md_ctx, md, NULL))) {
-        fprintf(stderr, "Digest init failed for algorithm: %s\n", algorithm);
-        EVP_MD_CTX_free(md_ctx);
-        RAISE_ERROR(BADARG_ATOM);
-    }
-
-    InteropFunctionResult result = interop_chardata_fold(data, hash_fold_fun, NULL, (void *) md_ctx);
-    if (result != InteropOk) {
-        fprintf(stderr, "Failed hashing input for algorithm: %s\n", algorithm);
-        RAISE_ERROR(BADARG_ATOM);
-    }
-
-    unsigned char digest[EVP_MAX_MD_SIZE];
-    unsigned int md_len;
-    if (!EVP_DigestFinal_ex(md_ctx, digest, &md_len)) {
-        fprintf(stderr, "Failed finishing hash for algorithm: %s\n", algorithm);
-        EVP_MD_CTX_free(md_ctx);
-        RAISE_ERROR(BADARG_ATOM);
-    }
-
-    EVP_MD_CTX_free(md_ctx);
-    if (UNLIKELY(memory_ensure_free(ctx, term_binary_heap_size(md_len)) != MEMORY_GC_OK)) {
-        RAISE_ERROR(OUT_OF_MEMORY_ATOM);
-    }
-
-    return term_from_literal_binary(digest, md_len, &ctx->heap, ctx->global);
-}
-
 static term nif_openssl_rand_bytes(Context *ctx, int argc, term argv[])
 {
     UNUSED(argc);
@@ -211,11 +97,6 @@ static term nif_openssl_random(Context *ctx, int argc, term argv[])
     return term_make_boxed_int(value, &ctx->heap);
 }
 
-static const struct Nif crypto_hash_nif =
-{
-    .base.type = NIFFunctionType,
-    .nif_ptr = nif_crypto_hash
-};
 static const struct Nif openssl_rand_bytes_nif =
 {
     .base.type = NIFFunctionType,
@@ -246,10 +127,6 @@ static const struct Nif atomvm_platform_nif =
 const struct Nif *platform_nifs_get_nif(const char *nifname)
 {
 #if defined ATOMVM_HAS_OPENSSL
-    if (strcmp("crypto:hash/2", nifname) == 0) {
-        TRACE("Resolved platform nif %s ...\n", nifname);
-        return &crypto_hash_nif;
-    }
     if (strcmp("atomvm:rand_bytes/1", nifname) == 0) {
         TRACE("Resolved platform nif %s ...\n", nifname);
         return &openssl_rand_bytes_nif;
@@ -269,6 +146,10 @@ const struct Nif *platform_nifs_get_nif(const char *nifname)
     }
     nif = otp_socket_nif_get_nif(nifname);
 #if defined ATOMVM_HAS_MBEDTLS
+    if (nif) {
+        return nif;
+    }
+    nif = otp_crypto_nif_get_nif(nifname);
     if (nif) {
         return nif;
     }
