+++ src/platforms/rp2040/src/lib/CMakeLists.txt
@@ -25,15 +25,18 @@ set(HEADER_FILES
     platform_defaultatoms.h
     platform_smp.h
     rp2040_sys.h
+    ../../../../libAtomVM/otp_crypto.h
 )
 
 set(SOURCE_FILES
     gpiodriver.c
     networkdriver.c
+    otp_crypto_platform.c
     platform_defaultatoms.c
     platform_nifs.c
     smp.c
     sys.c
+    ../../../../libAtomVM/otp_crypto.c
 )
 
 set(
@@ -55,6 +58,7 @@ target_link_libraries(
         hardware_rtc
         hardware_sync
         pico_float
+        pico_mbedtls
         pico_multicore
         pico_platform
         pico_runtime
@@ -91,8 +95,8 @@ if (PICO_CYW43_SUPPORTED)
             ../../../../libAtomVM/otp_ssl.h
             otp_net_lwip_raw.c
             otp_net_lwip_raw.h)
-    target_link_libraries(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC pico_cyw43_arch_lwip_threadsafe_background pico_lwip_sntp pico_mbedtls INTERFACE pan_lwip_dhserver)
+    target_link_libraries(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC pico_cyw43_arch_lwip_threadsafe_background pico_lwip_sntp INTERFACE pan_lwip_dhserver)
     target_link_options(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC "SHELL:-Wl,-u -Wl,networkregister_port_driver -Wl,-u -Wl,otp_socket_nif -Wl,-u -Wl,otp_net_nif -Wl,-u -Wl,otp_ssl_nif")
 endif()
 
-target_link_options(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC "SHELL:-Wl,-u -Wl,gpio_nif")
+target_link_options(libAtomVM${PLATFORM_LIB_SUFFIX} PUBLIC "SHELL:-Wl,-u -Wl,gpio_nif -Wl,-u -Wl,otp_crypto_nif")
