+++ src/platforms/rp2040/tests/test_erl_sources/CMakeLists.txt
@@ -24,11 +24,11 @@ ExternalProject_Add(HostAtomVM
     BUILD_COMMAND cmake --build . --target=atomvmlib --target=PackBEAM
 )
 
-function(compile_erlang module_name)
+function(compile_erlang module_name module_src_dir)
     add_custom_command(
         OUTPUT "${CMAKE_CURRENT_BINARY_DIR}/${module_name}.beam"
-        COMMAND erlc ${CMAKE_CURRENT_SOURCE_DIR}/${module_name}.erl
-        DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/${module_name}.erl"
+        COMMAND erlc ${CMAKE_CURRENT_SOURCE_DIR}/${module_src_dir}/${module_name}.erl
+        DEPENDS "${CMAKE_CURRENT_SOURCE_DIR}/${module_src_dir}/${module_name}.erl"
         COMMENT "Compiling ${module_name}.erl"
         WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
     )
@@ -36,8 +36,9 @@ function(compile_erlang module_name)
     set_property(DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}" APPEND PROPERTY ADDITIONAL_MAKE_CLEAN_FILES "${CMAKE_CURRENT_BINARY_DIR}/${module_name}.beam")
 endfunction()
 
-compile_erlang(test_clocks)
-compile_erlang(test_smp)
+compile_erlang(test_clocks "")
+compile_erlang(test_smp "")
+compile_erlang(test_crypto ../../../esp32/test/main/test_erl_sources/)
 
 add_custom_command(
     OUTPUT "${CMAKE_CURRENT_BINARY_DIR}/rp2040_test_modules.avm"
@@ -45,10 +46,12 @@ add_custom_command(
         HostAtomVM-prefix/src/HostAtomVM-build/libs/atomvmlib.avm
         test_clocks.beam
         test_smp.beam
+        test_crypto.beam
     DEPENDS
         HostAtomVM
         "${CMAKE_CURRENT_BINARY_DIR}/test_clocks.beam"
         "${CMAKE_CURRENT_BINARY_DIR}/test_smp.beam"
+        "${CMAKE_CURRENT_BINARY_DIR}/test_crypto.beam"
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
     VERBATIM
 )
